+++ apps/web/app/(app)/workspaces/[workspaceId]/unify/feedback-records/components/feedback-records-page-client.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "react-i18next";
 import type { FeedbackRecordData } from "@/modules/hub/types";
 import { PageContentWrapper } from "@/modules/ui/components/page-content-wrapper";
 import { PageHeader } from "@/modules/ui/components/page-header";
-import { UnifyConfigNavigation } from "../components/UnifyConfigNavigation";
+import { UnifyConfigNavigation } from "../../components/UnifyConfigNavigation";
 import { FeedbackRecordsTable } from "./feedback-records-table";
 
 interface FeedbackRecordsPageClientProps {
