+++ apps/web/locales/de-DE.json
@@ -3745,7 +3745,6 @@
       "source_name": "Quellenname",
       "source_type": "Quellentyp",
       "source_type_cannot_be_changed": "Quellentyp kann nicht geändert werden",
-      "sources": "Quellen",
       "status_error": "Fehler",
       "status_live_sync": "Live-Synchronisierung",
       "status_ready": "Bereit",
