+++ apps/web/locales/fr-FR.json
@@ -3745,7 +3745,6 @@
       "source_name": "Nom de la source",
       "source_type": "Type de source",
       "source_type_cannot_be_changed": "Le type de source ne peut pas être modifié",
-      "sources": "Sources",
       "status_error": "Erreur",
       "status_live_sync": "Synchronisation en direct",
       "status_ready": "Prêt",
