+++ apps/web/locales/ja-JP.json
@@ -3745,7 +3745,6 @@
       "source_name": "ソース名",
       "source_type": "ソースタイプ",
       "source_type_cannot_be_changed": "ソースタイプは変更できません",
-      "sources": "ソース",
       "status_error": "エラー",
       "status_live_sync": "リアルタイム同期",
       "status_ready": "準備完了",
