+++ apps/web/locales/sv-SE.json
@@ -3745,7 +3745,6 @@
       "source_name": "Källnamn",
       "source_type": "Källtyp",
       "source_type_cannot_be_changed": "Källtyp kan inte ändras",
-      "sources": "Källor",
       "status_error": "Fel",
       "status_live_sync": "Live sync",
       "status_ready": "Ready",
