+++ apps/web/locales/tr-TR.json
@@ -3745,7 +3745,6 @@
       "source_name": "Kaynak Adı",
       "source_type": "Kaynak Türü",
       "source_type_cannot_be_changed": "Kaynak türü değiştirilemez",
-      "sources": "Kaynaklar",
       "status_error": "Hata",
       "status_live_sync": "Live sync",
       "status_ready": "Ready",
